refactor: move DataComponentTypesInvoker to Accessor
