+++ src/main/resources/accessor.mixin.json
@@ -4,6 +4,7 @@
   "compatibilityLevel": "JAVA_16",
   "mixins": [
     "BeaconLevelAccessor",
+    "DataComponentTypesAccessor",
     "PlayerAbilitiesAccessor",
     "SerializingRegionBasedStorageAccessor",
     "ServerPlayerEntityFunctionAccessor",
