--- src/main/resources/invoker.mixin.json
@@ -1,12 +1,0 @@
-{
-  "required": true,
-  "package": "one.oktw.galaxy.mixin.invoker",
-  "compatibilityLevel": "JAVA_21",
-  "mixins": [
-    "DataComponentTypesInvoker"
-  ],
-  "client": [],
-  "injectors": {
-    "defaultRequire": 1
-  }
-}
